+++ ChangeLog.md
@@ -1,10 +1,18 @@
 For latest version of this document see [`ChangeLog.md on GitHub`
 ](https://github.com/trskop/socks-server/blob/main/ChangeLog.md).
 
-
 ### HEAD (unreleased changes)
 
 * TODO
+
+### 0.2.0.0 (unreleased changes)
+
+*   Support for Dhall standard 20.2.0 (Dhall library version 1.40.1).
+
+*   Command line options were renamed to be more consistent:
+
+    * `--typecheck` → `--config-typecheck`
+    * `--print-config-type` → `--config-print-type`
 
 ### 0.1.0.0